+++ src/templates/layout.js
@@ -13,7 +13,7 @@ import MobileNav from '../components/layout/MobileNav';
 const Layout = ({ children, path, seo }) => {
   // subscribes to value from ../utils/ThemeContext.js and
   // gives current theme to styled-components <ThemeProvider>
-  const theme = useContext(ThemeContext);
+  const { isLight, currentTheme } = useContext(ThemeContext);
   let [mobileNavIsOpen, toggleMobileNav] = useState(false);
 
   const mobileNavAction = () => {
@@ -31,8 +31,8 @@ const Layout = ({ children, path, seo }) => {
   });
 
   return (
-    <ThemeProvider theme={theme.currentTheme}>
-      <LayoutWrapper>
+    <ThemeProvider theme={currentTheme}>
+      <LayoutWrapper id="layoutWrapper" data-theme={isLight ? 'light' : 'dark'}>
         <Head seo={seo} />
         <div style={{ position: 'absolute', top: '0', zIndex: '1' }}>
           <Header />